Removed unused variables and fixed GCC warnings
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6259 212acab6-be3b-0410-9dea-997c60f758d6

+++ Contrib/NSISdl/asyncdns.cpp
@@ -27,7 +27,7 @@ JNL_AsyncDNS::~JNL_AsyncDNS()
   wait_for_thread_death();
 }
 
-unsigned long WINAPI JNL_AsyncDNS::_threadfunc(LPVOID _d)
+DWORD WINAPI JNL_AsyncDNS::_threadfunc(LPVOID _d)
 {
   JNL_AsyncDNS *_this=(JNL_AsyncDNS*)_d;
   struct hostent *hostentry;
